# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/booking-exceptions": {
      "get": {
        "operationId": "BookingExceptionsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "teacherCoreUserId",
            "required": false,
            "in": "query",
            "description": "ค่า sub ของอาจารย์ (ไม่ใช่ UUID เสมอไป)",
            "schema": {
              "maxLength": 64,
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-05",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-09",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BookingExceptionResponse"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายการเปิด/ปิดเวลาเฉพาะวันของอาจารย์",
        "tags": [
          "booking-exceptions"
        ]
      },
      "post": {
        "operationId": "BookingExceptionsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingExceptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BookingExceptionResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เปิดหรือปิดเวลาเฉพาะช่วงของตัวเอง",
        "tags": [
          "booking-exceptions"
        ]
      }
    },
    "/api/v1/booking-exceptions/{id}": {
      "delete": {
        "operationId": "BookingExceptionsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/DeletedModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ยกเลิกรายการเปิด/ปิดเวลา (ย้อนกลับไปใช้เวลาทำการรายสัปดาห์)",
        "tags": [
          "booking-exceptions"
        ]
      }
    },
    "/api/v1/bookings": {
      "get": {
        "operationId": "BookingsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "CONFIRMED",
                "CANCELLED",
                "COMPLETED"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-05",
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-10-09",
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "default": "startsAt",
              "type": "string",
              "enum": [
                "startsAt",
                "-startsAt",
                "-createdAt"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/BookingResponse"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMetaModel"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "คิวนัดหมายของฉัน (อาจารย์เห็นคิวที่ตัวเองดูแล นักศึกษาเห็นคิวของตัวเอง)",
        "tags": [
          "bookings"
        ]
      },
      "post": {
        "operationId": "BookingsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBookingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BookingResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "อาจารย์สร้างนัดหมายให้นักศึกษาในคิวของตัวเอง",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{id}": {
      "get": {
        "operationId": "BookingsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BookingResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ดูนัดหมายรายการเดียว",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{id}/cancel": {
      "patch": {
        "operationId": "BookingsController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BookingResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ยกเลิกนัดหมาย (ช่วงเวลากลับไปว่าง)",
        "tags": [
          "bookings"
        ]
      }
    },
    "/api/v1/bookings/{id}/complete": {
      "patch": {
        "operationId": "BookingsController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/BookingResponse"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ทำเครื่องหมายว่านัดหมายเสร็จสิ้น",
        "tags": [
          "bookings"
        ]
      }
    },
...
  },
  "info": {
    "title": "Q-Teacher API",
    "description": "ระบบนัดหมายเข้าพบอาจารย์ (CSMJU2030) — ตอบเป็น { success, data[, meta] } หรือ { success: false, error } ตามมาตรฐาน api-conventions.md · ต้องมี token ของ Core Hub ทุก endpoint ยกเว้น /api/health",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csmju_q_teacher_access_token"
      }
    },
    "schemas": {
      "PaginationMetaModel": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 42
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "totalPages": {
            "type": "number",
            "example": 3
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "totalPages"
        ]
      },
...
      "DeletedModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "deleted": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "id",
          "deleted"
        ]
      },
      "BookingExceptionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "teacherCoreUserId": {
            "type": "string"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ]
          },
          "reason": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "teacherCoreUserId",
          "startsAt",
          "endsAt",
          "status",
          "reason",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBookingExceptionDto": {
        "type": "object",
        "properties": {
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-05T12:00:00+07:00"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-05T13:00:00+07:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "AVAILABLE",
              "UNAVAILABLE"
            ],
            "description": "AVAILABLE = เปิดเวลานี้ให้จอง · UNAVAILABLE = ปิดเวลานี้"
          },
          "reason": {
            "type": "string",
            "maxLength": 200
          }
        },
        "required": [
          "startsAt",
          "endsAt",
          "status"
        ]
      },
      "BookingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "teacherCoreUserId": {
            "type": "string",
            "description": "ค่า sub ของอาจารย์"
          },
          "teacherPersonCode": {
            "type": "string",
            "nullable": true,
            "description": "รหัสบุคลากรจาก Core Hub (แสดงแทนชื่อ)"
          },
          "studentCoreUserId": {
            "type": "string",
            "description": "ค่า sub ของนักศึกษา"
          },
          "studentPersonCode": {
            "type": "string",
            "nullable": true,
            "description": "รหัสนักศึกษาจาก Core Hub (แสดงแทนชื่อ)"
          },
          "studentFullNameTh": {
            "type": "string",
            "nullable": true,
            "description": "ชื่อนักศึกษา อ่านจาก Core Hub ตอนเปิดดูรายการเดียวและเฉพาะผู้ที่มีสิทธิ์อ่านข้อมูลบุคคล — ไม่เก็บในฐานข้อมูล"
          },
          "startsAt": {
            "type": "string",
            "format": "date-time"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time"
          },
          "topic": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "CONFIRMED",
              "CANCELLED",
              "COMPLETED"
            ]
          },
          "unreadCount": {
            "type": "number",
            "description": "จำนวนข้อความในคิวนี้ที่อีกฝ่ายส่งมาและยังไม่ได้อ่าน"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "teacherCoreUserId",
          "teacherPersonCode",
          "studentCoreUserId",
          "studentPersonCode",
          "studentFullNameTh",
          "startsAt",
          "endsAt",
          "topic",
          "status",
          "unreadCount",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateBookingDto": {
        "type": "object",
        "properties": {
          "studentPersonCode": {
            "type": "string",
            "example": "6704101312",
            "maxLength": 50
          },
          "startsAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-12T09:00:00+07:00"
          },
          "endsAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-10-12T09:30:00+07:00"
          },
          "topic": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "example": "ปรึกษาโครงงานพิเศษ"
          }
        },
        "required": [
          "studentPersonCode",
          "startsAt",
          "endsAt",
          "topic"
        ]
      },
...
    }
  }
}